requirements.txt
python-dotenv==0.19.*
requests==2.26.*
pip==21.3.1
wheel==0.37.0
tornado==6.1
pytz==2021.3
certifi==2021.10.*
tzlocal==4.1
urllib3==1.26.*
six==1.16.0
idna==3.3
APScheduler==3.6.*
cachetools==4.2.2
setuptools==58.4.0